[refactor/#176] 수정삭제 팝오버 리팩토링
