+++ src/components/Sollect/SollectPhoto.tsx
@@ -5,38 +5,26 @@ import SollectMark from './SollectMark';
 import kebabWhite from '../../assets/kebabWhite.svg';
 import EditDeletePopover from '../global/EditDeletePopover';
 import { deleteSollect } from '../../api/sollectApi';
-import Modal from '../global/Modal';
 
-type Props = SollectPhotoProps &{
-  isMine?:boolean;
-}
+type Props = SollectPhotoProps & {
+  isMine?: boolean;
+};
 
 const SollectPhoto: React.FC<Props> = ({ sollect, isMine }) => {
   const navigate = useNavigate();
   const [marked, setMarked] = useState(sollect.isMarked);
   const [showMenu, setShowMenu] = useState(false);
-    const [showDeleteModal, setShowDeleteModal] = useState(false);
 
   const handleClick = () => {
     navigate(`/sollect/${sollect.sollectId}`);
   };
 
   // 쏠마크 마이페이지에서 사용
-    const clickDeleteModal = () => {
-    setShowDeleteModal(true);
+  const funcDelete = async () => {
+    await deleteSollect(Number(sollect.sollectId));
+    setShowMenu(false);
+    location.reload();
   };
-
-    const onLeftClick = () => {
-      setShowDeleteModal(false);
-    };
-  
-    const onRightClick = async () => {
-      setShowDeleteModal(false);
-      await deleteSollect(Number(sollect.sollectId));
-      setShowMenu(false);
-      location.reload();
-    };
-  
 
   return (
     <div
@@ -63,9 +51,7 @@ const SollectPhoto: React.FC<Props> = ({ sollect, isMine }) => {
               className='w-24 h-24'
               onClick={() => setShowMenu(!showMenu)}
             />
-            {showMenu && (
-              <EditDeletePopover clickDeleteModal={clickDeleteModal} />
-            )}
+            {showMenu && <EditDeletePopover funcDelete={funcDelete} />}
           </div>
         ) : (
           <SollectMark
@@ -74,17 +60,6 @@ const SollectPhoto: React.FC<Props> = ({ sollect, isMine }) => {
             id={sollect.sollectId}
           />
         )}
-
-        {showDeleteModal && (
-            <Modal
-              title='정말 삭제하시겠습니까?'
-              subtitle='삭제하면 복구할 수 없어요!'
-              leftText='취소'
-              rightText='삭제'
-              onLeftClick={onLeftClick}
-              onRightClick={onRightClick}
-            />
-          )}
       </div>
 
       <div className='p-20 z-1 text-white'>
